fix(re-audit-v2): remediate actionable findings (timeouts, canonical signature, numeric recompute)
Implements the actionable findings from re-auditoría v2 (docs/audits/
re-audit-v2-delta.md). No P0/P1 were raised; these harden P2/P3 items and
document intentional designs. Verified against PG16 real: full suite (558
tests incl. 4 new), migrate ×2, verify-ledger-invariants.sql →
FLUVIA_INVARIANTS_OK, and all 6 runbook drills PASS.

- V2-R1 (P2) RESUELTO — withTenantTransaction fija SET LOCAL statement_timeout
  (30s), lock_timeout (15s) e idle_in_transaction_session_timeout (60s),
  sobreescribibles por transacción, inyectados junto al tenant en un solo
  set_config parametrizado. Evita agotamiento de pool por query/lock/tx colgada.
  4 tests nuevos (packages/db/test/tx-timeouts.test.ts) incl. abort real 57014.
- V2-R2 (P2 latente) RESUELTO — entriesSignature usa JSON.stringify de tuplas
  posicionales ordenadas por forma canónica; imposible colisión por separador.
  Se recomputa en ambos lados y nunca se persiste ⇒ sin migración.
- V2-R4 (P2) RESUELTO — verifyProjection/rebuildProjection e invariantes SQL
  recomputan en ::numeric (SUM(bigint) ya es numeric; evita cast intermedio a
  bigint que abortaría en el borde). numeric::text == bigint::text para enteros.
- V2-R6 (P3) aclarado — comentarios distinguen init-de-acumulador y cero
  defensivo de un fallback silencioso en applyProjectionDeltas.
- V2-R5/R7/N3 (P3) DOCUMENTADO — sesgo determinista de Money.allocate; 40001
  forward-looking bajo READ COMMITTED; lectura best-effort de refund.execute.

Registro de estado por hallazgo (incl. V2-N1 verificado no-explotable, V2-N4
guard de dos-humanos DB-enforced, V2-R3/N2 diferidos) en
docs/audits/audit-closure-register-v1.md.

+++ scripts/verify-ledger-invariants.sql
@@ -67,14 +67,18 @@ BEGIN
   -- 6. Toda proyeccion EXISTENTE == recomputo desde asientos (por bucket,
   --    con normal_side). Una cuenta sin fila de proyeccion no es corrupcion
   --    silenciosa (getBalance falla visible; rebuild la materializa).
+  -- V2-R4 (re-auditoria v2): el recomputo se mantiene en `numeric`. `SUM(bigint)`
+  -- ya devuelve numeric (sin overflow), pero un `::bigint` explicito abortaria
+  -- con "bigint out of range" en el borde; en numeric la comparacion contra la
+  -- proyeccion (bigint) sigue siendo exacta y jamas lanza por un cast.
   WITH recomputed AS (
     SELECT e.account_id,
       COALESCE(SUM(CASE WHEN e.bucket = 'available'
         THEN CASE WHEN e.direction = a.normal_side THEN e.amount ELSE -e.amount END
-        ELSE 0 END), 0)::bigint AS available,
+        ELSE 0 END), 0)::numeric AS available,
       COALESCE(SUM(CASE WHEN e.bucket = 'pending'
         THEN CASE WHEN e.direction = a.normal_side THEN e.amount ELSE -e.amount END
-        ELSE 0 END), 0)::bigint AS pending
+        ELSE 0 END), 0)::numeric AS pending
     FROM ledger_entries e
     JOIN ledger_accounts a ON a.id = e.account_id
     GROUP BY e.account_id
